fix: XHCI babble error on E4870
The filter size is UINT32 instead of UINT16. USB NIC returns two more
bytes back to host and cause babble error.

Tested-by: Ashish Singhal <ashishsingha@nvidia.com>

+++ Silicon/NVIDIA/Drivers/BmcNic/UsbRndisDxe/UsbRndisDxe.h
@@ -90,7 +90,7 @@ typedef struct {
   UINT32               MediaStatus;
   UINT32               LinkSpeed;
   UINT32               MaxFrameSize;
-  UINT16               Filter;
+  UINT32               Filter;
   UINT32               Medium;
   UINT32               MaxPacketsPerTransfer;
   UINT32               MaxTransferSize;
